+++ api/lib/application/authentication/oidc/index.js
@@ -35,32 +35,6 @@ const register = async function (server) {
     ...adminRoutes,
     {
       method: 'POST',
-      path: '/api/oidc/token',
-      config: {
-        auth: false,
-        validate: {
-          payload: Joi.object({
-            data: {
-              attributes: {
-                identity_provider: Joi.string().required(),
-                code: Joi.string().required(),
-                redirect_uri: Joi.string().required(),
-                state: Joi.string().required(),
-                audience: Joi.string().valid('app', 'admin').optional(),
-              },
-            },
-          }),
-        },
-        handler: oidcController.authenticateUser,
-        notes: [
-          "- Cette route permet de récupérer un token pour un utilisateur provenant d'un partenaire.\n" +
-            "- Elle retournera également un access token Pix correspondant à l'utilisateur.",
-        ],
-        tags: ['api', 'authentication', 'oidc'],
-      },
-    },
-    {
-      method: 'POST',
       path: '/api/oidc/user/check-reconciliation',
       config: {
         auth: false,
